+++ index.ts
@@ -1,6 +1,2 @@
 export { default } from "./src/pi.js";
-export * from "./src/schema.js";
-export * from "./src/outbox.js";
-export * from "./src/cognee.js";
-export * from "./src/multica.js";
-export * from "./src/runtime.js";
+export * from "./src/core.js";